silae_planning/requirements.txt
google-auth~=2.41.1
google-auth-oauthlib~=1.2.2
google-auth-httplib2~=0.2.0
google-api-python-client~=2.185.0
pytz~=2023.3
requests~=2.31.0
python-dotenv~=1.0.0